#778 Moved MoveGizmo into a file of its own.
